Move auction pricing and its reference anchor into the Auction algorithm
The clearing-price computation lived on Matcher, and the book computed
a price then wrapped a freshly-constructed Auction around it - so the
auction's own policy was split across two collaborators that both had
to agree on it.

TryComputeAuctionPrice, its tie-break and SumRemaining all move to
Auction as TryComputeClearingPrice. The reference anchor moves with
them: it was documented as feeding the auction tie-break and nothing
else, so Auction now owns it behind OnTrade/OnSessionChange hooks, the
same way each IPriceRestriction owns its anchor. InMemoryOrderBook
drops _auctionReferencePriceTicks and forwards to those hooks instead.

IMatchingAlgorithm gains TryBegin, called once by Matcher.Run to let an
algorithm derive run-scoped state from the book - where Auction strikes
its clearing price, fixing it for the print rather than re-deriving it
per trade from a book those trades are consuming. Returning false means
nothing to match, so an uncrossing pass over an uncrossed book is a
no-op the caller no longer has to guard: OpenMarket just hands Match
the long-lived _auction instance, and TryGetIndicativeAuctionPrice asks
that same instance for a price without committing to a print.

No behavior change.
